prescript glasses for menu
+ optimize skin creation
