+++ game/core/src/com/shardbytes/ripsafarik/game/MainGame.kt
@@ -3,6 +3,7 @@ package com.shardbytes.ripsafarik.game
 import com.badlogic.gdx.Game
 import com.badlogic.gdx.assets.AssetManager
 import com.shardbytes.ripsafarik.assets.Animations
+import com.shardbytes.ripsafarik.assets.Skin
 import com.shardbytes.ripsafarik.assets.Textures
 import com.shardbytes.ripsafarik.screens.MenuScreen
 
@@ -18,6 +19,7 @@ object MainGame : Game() {
 	override fun create() {
 		Textures.manager = assetManager
 		Textures.loadAll() //Load all textures
+		Skin.create() //Create skin for GUI elements
 
 		//Load sounds and other assets if required
 		assetManager.finishLoading()
