+++ game/core/src/com/shardbytes/ripsafarik/screens/PauseScreen.kt
@@ -3,23 +3,17 @@ package com.shardbytes.ripsafarik.screens
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.Screen
 import com.badlogic.gdx.graphics.*
-import com.badlogic.gdx.graphics.g2d.BitmapFont
 import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.Stage
 import com.badlogic.gdx.scenes.scene2d.ui.Label
-import com.badlogic.gdx.scenes.scene2d.ui.ScrollPane
-import com.badlogic.gdx.scenes.scene2d.ui.Skin
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener
 import com.badlogic.gdx.utils.Scaling
 import com.badlogic.gdx.utils.viewport.ScalingViewport
+import com.shardbytes.ripsafarik.assets.Skin.skin
 import com.shardbytes.ripsafarik.game.MainGame
 import com.shardbytes.ripsafarik.components.input.InputCore
-import com.shardbytes.ripsafarik.entity.Player
 import com.shardbytes.ripsafarik.tools.SaveManager
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonConfiguration
-import kotlinx.serialization.stringify
 
 //TODO: Šimon plz finish
 //no u
@@ -35,24 +29,14 @@ class PauseScreen : Screen {
 
 	)
 
-	private var buttonHeight = uiStage.viewport.screenHeight / 9f
-	private var buttonWidth = uiStage.viewport.screenWidth / 5f
-	private var centeredButtonX = uiStage.viewport.screenWidth / 3 / 2 - buttonWidth / 2
-	private var centeredButtonY = uiStage.viewport.screenHeight / 2 - buttonHeight / 2
-
-	private var label: Label
-	private var loadButton: TextButton
-	private var saveButton: TextButton
-	private var exitButton: TextButton
+	private var label = Label("Paused", skin)
+	private var loadButton = TextButton("Resume", skin)
+	private var saveButton = TextButton("Save", skin)
+	private var exitButton = TextButton("Exit", skin)
 
 	init {
 		Gdx.input.inputProcessor = uiStage
-		val skin = createSkin()
-
-		loadButton = TextButton("Resume", skin)
-		loadButton.height = buttonHeight
-		loadButton.width = buttonWidth
-		loadButton.setPosition(centeredButtonX, centeredButtonY)
+		
 		loadButton.addListener(object : ChangeListener() {
 			override fun changed(event: ChangeEvent, actor: Actor) {
 				Gdx.input.inputProcessor = InputCore.reset()
@@ -62,11 +46,7 @@ class PauseScreen : Screen {
 			}
 
 		})
-
-		saveButton = TextButton("Save", skin)
-		saveButton.height = buttonHeight
-		saveButton.width = buttonWidth
-		saveButton.setPosition(centeredButtonX, centeredButtonY - buttonHeight * 1.25f)
+		
 		saveButton.addListener(object : ChangeListener() {
 			override fun changed(event: ChangeEvent, actor: Actor) {
 				SaveManager.save()
@@ -77,11 +57,7 @@ class PauseScreen : Screen {
 			}
 
 		})
-
-		exitButton = TextButton("Exit", skin)
-		exitButton.height = buttonHeight
-		exitButton.width = buttonWidth
-		exitButton.setPosition(centeredButtonX, centeredButtonY - buttonHeight * 2.5f)
+		
 		exitButton.addListener(object : ChangeListener() {
 			override fun changed(event: ChangeEvent, actor: Actor) {
 				Gdx.app.exit()
@@ -90,47 +66,34 @@ class PauseScreen : Screen {
 
 		})
 
-		label = Label("Paused", skin)
-		label.setPosition(Gdx.graphics.width / 3 / 2 - label.width / 2, Gdx.graphics.height / 2 - label.height / 2 + Gdx.graphics.height / 5)
-
 		uiStage.addActor(loadButton)
 		uiStage.addActor(saveButton)
 		uiStage.addActor(exitButton)
 		uiStage.addActor(label)
+		
+		setPositionOfUIElements()
 
 	}
-
-	private fun createSkin(): Skin {
-		val skin = Skin()
-		val pixmap = Pixmap(1, 1, Pixmap.Format.RGBA8888)
-		pixmap.setColor(Color.WHITE)
-		pixmap.fill()
-
-		skin.add("white", Texture(pixmap))
-		skin.add("defaultFont", BitmapFont())
-		skin.add("whiteColour", Color.WHITE)
-
-		val style = TextButton.TextButtonStyle()
-		style.font = skin.getFont("defaultFont")
-		style.up = skin.newDrawable("white", Color.DARK_GRAY)
-		style.down = skin.newDrawable("white", Color.DARK_GRAY)
-		style.checked = skin.newDrawable("white", Color.BLUE)
-		style.over = skin.newDrawable("white", Color.LIGHT_GRAY)
-		skin.add("default", style)
-
-		val style1 = Label.LabelStyle()
-		style1.font = skin.getFont("defaultFont")
-		style1.fontColor = skin.getColor("whiteColour")
-		skin.add("default", style1)
-		val style2 = ScrollPane.ScrollPaneStyle()
-
-		style2.background = skin.newDrawable("white", Color.DARK_GRAY)
-		style2.corner = skin.newDrawable("white", Color.CYAN)
-		style2.vScrollKnob = skin.newDrawable("white", Color.WHITE)
-		skin.add("default", style2)
-
-		return skin
-
+	
+	private fun setPositionOfUIElements() {
+		val buttonHeight = uiStage.viewport.screenHeight / 9f
+		val buttonWidth = uiStage.viewport.screenWidth / 5f
+		val centeredButtonX = uiStage.viewport.screenWidth / 3 / 2 - buttonWidth / 2
+		val centeredButtonY = uiStage.viewport.screenHeight / 2 - buttonHeight / 2
+		
+		loadButton.height = buttonHeight
+		loadButton.width = buttonWidth
+		loadButton.setPosition(centeredButtonX, centeredButtonY)
+		
+		saveButton.height = buttonHeight
+		saveButton.width = buttonWidth
+		saveButton.setPosition(centeredButtonX, centeredButtonY - buttonHeight * 1.25f)
+		
+		exitButton.height = buttonHeight
+		exitButton.width = buttonWidth
+		exitButton.setPosition(centeredButtonX, centeredButtonY - buttonHeight * 2.5f)
+		
+		label.setPosition(Gdx.graphics.width / 3 / 2 - label.width / 2, Gdx.graphics.height / 2 - label.height / 2 + Gdx.graphics.height / 5)
 	}
 
 	override fun show() {}
@@ -151,16 +114,7 @@ class PauseScreen : Screen {
 
 	override fun resize(width: Int, height: Int) {
 		uiStage.viewport.update(width, height)
-
-		buttonHeight = height / 9f
-		buttonWidth = width / 5f
-		centeredButtonX = width / 3 / 2 - buttonWidth / 2
-		centeredButtonY = height / 2 - buttonHeight / 2
-
-		loadButton.setPosition(centeredButtonX, centeredButtonY)
-		exitButton.setPosition(centeredButtonX, centeredButtonY - buttonHeight * 2.5f)
-
-		label.setPosition(Gdx.graphics.width / 3 / 2 - label.width / 2, Gdx.graphics.height / 2 - label.height / 2 + Gdx.graphics.height / 5)
+		setPositionOfUIElements()
 
 	}
